Move copy_to_current to _copy
